some more fixes
+ aie_arcopen() now accepts strings and handles file opening and closing by itself
+ aie_ARCFORMATS_COUNT is now of type size_t
+ aie_kmarctable() does not leak anymore

+++ include/aie_archive-formats.h
@@ -4,4 +4,4 @@
 
 extern const aie_ArcFormat const * aie_arcformats[];
 
-extern const int aie_ARCFORMATS_COUNT;
+extern const size_t aie_ARCFORMATS_COUNT;
