+++ src/archive-formats.c
@@ -7,4 +7,4 @@ const aie_ArcFormat const * aie_arcformats[] =
   &kid_engine_link
 };
 
-const int aie_ARCFORMATS_COUNT = sizeof aie_arcformats / sizeof (void *);
+const size_t aie_ARCFORMATS_COUNT = sizeof aie_arcformats / sizeof (void *);
